Story 9-5: extract home_indicators.rs (Task 1, pure move)
Move IndicatorFilter, IndicatorTag, parse_indicator_filter,
build_indicator_tags + their direct unit tests out of home.rs into a
sibling module. Net effect: home.rs drops from 1987 → 1771 LOC,
reclaiming Foundation Rule #12 headroom for stories 9-5 through 9-7
which each add a new indicator variant + parser arm + helper block.

This is a pure code move — zero behavior change, zero new tests, no
new logic. Story 9-5 substance lands in subsequent commits (Tasks 2+).

The IndicatorTag struct stays `pub` so the publicly-reachable
HomeTemplate.indicator_tags field type matches its visibility; the
enum + functions become `pub(crate)` since their only callers are
inside the crate.

Sprint-status updated 9-5 → in-progress (only that line + last_updated).
