+++ _bmad-output/implementation-artifacts/sprint-status.yaml
@@ -39,7 +39,7 @@
 # The epics document contains FR/NFR/AR/UX-DR assignments per epic.
 
 generated: 2026-03-29
-last_updated: 2026-05-02 # Story 9-4 done — code review complete (post-merge of PR #114). 3 parallel reviewers (Blind/Edge/Auditor), ~27 findings: 0 decision-needed, 4 patches applied (1 Medium UX — active filter at zero count strands user → helper + macro both fixed to always emit active-state pill; 3 Low coverage gaps — AC1 placement test + id DESC tiebreak test + FilterTag 4th-corner matrix test), 0 deferred, 16+ dismissed. Final 671 lib + 4 dashboard_unshelved integration tests, all green; clippy clean. Earlier today: 9-4 dev-story complete. 9 tasks shipped: volume::count_unshelved + list_unshelved + UnshelvedVolumeRow struct (1 INNER JOIN round-trip with primary contributor subquery, mirroring 9-2's list_recent_active projection); IndicatorFilter enum + parse_indicator_filter sibling parser (':'-presence guard for namespace disambiguation, !is_empty() guard added beyond spec to avoid spurious WARN logs); IndicatorTag view-model + build_indicator_tags helper (zero-count rule baked in); home handler wiring with anonymous short-circuit BEFORE count query (AC2 no-leak, two-layer defense) + AC7 single-active-filter precedence (indicator wins, query/legacy filter cleared, has_filter=false naturally short-circuits HTMX search-fragment branch); FilterTag Askama macro (default vs active state, count>0 internal guard); #what-needs-attention section at top of dashboard (above #collection-glance); #unshelved-list mutually exclusive with #recent-additions in same DOM position (AC6); EN+FR i18n; 16 new lib tests (6 parser + 3 build_indicator_tags + 6 handler render + 1 macro guard) + 3 sqlx integration tests. 668 lib + 3 integration tests, all green; clippy clean; sqlx cache unchanged. Earlier 2026-05-02: spec creation; earlier 2026-05-01: Story 9-3 done Key decisions: schema column is volumes.location_id NOT storage_location_id (spec text drift); IndicatorFilter enum + parse_indicator_filter sibling parser (closed enum, ':'-presence guard avoids false positives on legacy genre:/state: filters); plain <a href> over HTMX boost in v1 (deviation from spec, documented with rationale; HTMX polish deferred to follow-up story); UnshelvedVolumeRow new struct (NOT SearchResult reuse — title-centric vs volume-centric); macro at templates/components/filter_tag.html mirroring cover.html shape; #what-needs-attention placed at TOP of dashboard (above #collection-glance); #unshelved-list mutually exclusive with #recent-additions in same DOM position; anonymous role short-circuits BEFORE count query (no leak, no DB load); 14 new tests including 4-state FilterTag macro matrix + AC2 anonymous-no-leak + AC6 mutual-exclusion + AC11c parser matrix. Earlier 2026-05-01: Story 9-3 done (code review post-merge complete, PRs #110 + #113 merged); 9-2 done; spec creation 9-2/9-3: single round-trip GROUP BY, INNER JOIN naturally excludes empty/soft-deleted genres (AC4/AC5), `<progress>` element for CSP-clean variable-width bar (UX-DR24 tokens via @theme vars), locale-aware percentage formatting (FR uses NBSP + comma decimal), row-link targets `/?filter=genre:<id>` (existing route, NOT the spec's `/catalog?genre=` which doesn't exist — same convention as 9-1's volume-link deviation), 5 handler render tests including document-order + anonymous/librarian byte-identity locks, format_percent helper in src/utils.rs with NBSP regression test. Earlier 2026-04-30: Story 9-2 review complete — review → done; Story 9-1 done + 2 GH issues filed (#103/#104), Epic 9 sprint planning (PR #101), Epic 8 close.
+last_updated: 2026-05-02 # Story 9-5 dev-story started → in-progress. Spec created earlier today. Story 9-4 done — code review complete (post-merge of PR #114). 3 parallel reviewers (Blind/Edge/Auditor), ~27 findings: 0 decision-needed, 4 patches applied (1 Medium UX — active filter at zero count strands user → helper + macro both fixed to always emit active-state pill; 3 Low coverage gaps — AC1 placement test + id DESC tiebreak test + FilterTag 4th-corner matrix test), 0 deferred, 16+ dismissed. Final 671 lib + 4 dashboard_unshelved integration tests, all green; clippy clean. Earlier today: 9-4 dev-story complete. 9 tasks shipped: volume::count_unshelved + list_unshelved + UnshelvedVolumeRow struct (1 INNER JOIN round-trip with primary contributor subquery, mirroring 9-2's list_recent_active projection); IndicatorFilter enum + parse_indicator_filter sibling parser (':'-presence guard for namespace disambiguation, !is_empty() guard added beyond spec to avoid spurious WARN logs); IndicatorTag view-model + build_indicator_tags helper (zero-count rule baked in); home handler wiring with anonymous short-circuit BEFORE count query (AC2 no-leak, two-layer defense) + AC7 single-active-filter precedence (indicator wins, query/legacy filter cleared, has_filter=false naturally short-circuits HTMX search-fragment branch); FilterTag Askama macro (default vs active state, count>0 internal guard); #what-needs-attention section at top of dashboard (above #collection-glance); #unshelved-list mutually exclusive with #recent-additions in same DOM position (AC6); EN+FR i18n; 16 new lib tests (6 parser + 3 build_indicator_tags + 6 handler render + 1 macro guard) + 3 sqlx integration tests. 668 lib + 3 integration tests, all green; clippy clean; sqlx cache unchanged. Earlier 2026-05-02: spec creation; earlier 2026-05-01: Story 9-3 done Key decisions: schema column is volumes.location_id NOT storage_location_id (spec text drift); IndicatorFilter enum + parse_indicator_filter sibling parser (closed enum, ':'-presence guard avoids false positives on legacy genre:/state: filters); plain <a href> over HTMX boost in v1 (deviation from spec, documented with rationale; HTMX polish deferred to follow-up story); UnshelvedVolumeRow new struct (NOT SearchResult reuse — title-centric vs volume-centric); macro at templates/components/filter_tag.html mirroring cover.html shape; #what-needs-attention placed at TOP of dashboard (above #collection-glance); #unshelved-list mutually exclusive with #recent-additions in same DOM position; anonymous role short-circuits BEFORE count query (no leak, no DB load); 14 new tests including 4-state FilterTag macro matrix + AC2 anonymous-no-leak + AC6 mutual-exclusion + AC11c parser matrix. Earlier 2026-05-01: Story 9-3 done (code review post-merge complete, PRs #110 + #113 merged); 9-2 done; spec creation 9-2/9-3: single round-trip GROUP BY, INNER JOIN naturally excludes empty/soft-deleted genres (AC4/AC5), `<progress>` element for CSP-clean variable-width bar (UX-DR24 tokens via @theme vars), locale-aware percentage formatting (FR uses NBSP + comma decimal), row-link targets `/?filter=genre:<id>` (existing route, NOT the spec's `/catalog?genre=` which doesn't exist — same convention as 9-1's volume-link deviation), 5 handler render tests including document-order + anonymous/librarian byte-identity locks, format_percent helper in src/utils.rs with NBSP regression test. Earlier 2026-04-30: Story 9-2 review complete — review → done; Story 9-1 done + 2 GH issues filed (#103/#104), Epic 9 sprint planning (PR #101), Epic 8 close.
 project: mybibli
 project_key: NOKEY
 tracking_system: file-system
@@ -167,7 +167,7 @@ development_status:
   9-2-dashboard-recent-additions: done
   9-3-dashboard-stats-by-genre: done
   9-4-filtertag-and-unshelved-indicator: done
-  9-5-overdue-loans-indicator: backlog
+  9-5-overdue-loans-indicator: in-progress
   9-6-series-with-gaps-indicator: backlog
   9-7-recent-activity-indicators: backlog
   9-8-loan-status-role-aware: backlog
